Update DBRS memcached centos8 dockerfile
As part of the DBRSv0.2.1 release, it includes a version bump of memcached version
to 1.6.2 wich contains an important security fix. ClearLinux's dockerfile was not
updated as it pulls memcached from a bundle and we do not manage the versioning
under this approach, but as today, the Clear Linux OS contains a memcached 1.6.2+
version.

Signed-off-by: Gabriel Briones Sayeg <gabriel.briones.sayeg@intel.com>